Refactor [#45] : MonthlyItemSales 변경
